+++ aws-lambda.cjs
@@ -1,0 +1,7 @@
+'use strict'
+
+exports = module.exports = {}
+
+exports.createApi = require('./commonjs/aws-lambda/api.js').createApi
+exports.updateRoutes = require('./commonjs/aws-lambda/api.js').updateApi
+exports.deploy = require('./commonjs/aws-lambda/deploy.js').default