+++ aws-lambda.js
@@ -1,7 +1,2 @@
-'use strict'
-
-exports = module.exports = {}
-
-exports.createApi = require('./commonjs/aws-lambda/api').createApi
-exports.updateRoutes = require('./commonjs/aws-lambda/api').updateApi
-exports.deploy = require('./commonjs/aws-lambda/deploy').default
+export { createApi, updateApi as updateRoutes } from './modules/aws-lambda/api.js'
+export { default as deploy } from './modules/aws-lambda/deploy.js'