+++ index.commonjs.js
@@ -1,5 +1,0 @@
-'use strict'
-
-exports = module.exports = {}
-
-exports.run = require('./commonjs/run/run').default