+++ source/code/pieces/createErrorResponse.js
@@ -1,12 +1,15 @@
 function $createErrorResponse(error) {
-  const errorMessage = error.httpStatusCode ? error.message : 'Error'
+  const errorObject = {
+    message: error.message || 'Error',
+    ...error.data
+  }
   const statusCode = error.httpStatusCode ? error.httpStatusCode : 500
   return {
     // isBase64Encoded: false,
     statusCode,
     headers: $headers(error),
     body: JSON.stringify({
-      errorMessage,
+      error: errorObject,
       statusCode
     })
   }