+++ source/dirname.js
@@ -1,0 +1,9 @@
+let dirname
+
+export function setDirname(_dirname) {
+	dirname = _dirname
+}
+
+export function getDirname() {
+	return dirname
+}